add codes

+++ src/common/btree/btree_array_list.cc
@@ -5,8 +5,8 @@
 
 namespace oceanbase {
 namespace common {
-const int32_t BtreeArrayList::NODE_SIZE = (CONST_NODE_OBJECT_COUNT* sizeof(BtreeKeyValuePair) + sizeof(BtreeNode));
-const int32_t BtreeArrayList::NODE_COUNT_PRE_PAGE = (NODE_SIZE / sizeof(void*) - 1);
+const int32_t BtreeArrayList::NODE_SIZE = (int32_t)(CONST_NODE_OBJECT_COUNT* sizeof(BtreeKeyValuePair) + sizeof(BtreeNode));
+const int32_t BtreeArrayList::NODE_COUNT_PRE_PAGE = (int32_t)(NODE_SIZE / sizeof(void*) - 1);
 
 BtreeArrayList::BtreeArrayList() {
   init(NODE_SIZE);
